+++ src/day2.rs
@@ -1,6 +1,6 @@
 use super::input;
 
-pub fn problem_1(data: &str) -> i32 {
+pub fn problem_1(data: &str) -> i64 {
     let data: input::File = input::read_file(data);
 
     data.rows()
@@ -16,7 +16,7 @@ pub fn problem_1(data: &str) -> i32 {
         .unwrap()
 }
 
-pub fn problem_2(data: &str) -> i32 {
+pub fn problem_2(data: &str) -> i64 {
     let data: input::File = input::read_file(data);
 
     data.rows()
@@ -25,7 +25,7 @@ pub fn problem_2(data: &str) -> i32 {
             (0..row.len() + 1).any(|element_to_skip| {
                 // If element_to_skip == row.len() then no element is skipped.
                 // This ensures we still accept a row that is valid with no skipped elements.
-                let row: Vec<i32> = Skipper::new(row, element_to_skip).collect();
+                let row: Vec<i64> = Skipper::new(row, element_to_skip).collect();
                 (0..row.len() - 1).all(|i| {
                     let diff = (row[i] - row[i + 1]).abs();
                     diff >= 1 && diff <= 3 && row[i].cmp(&row[i + 1]) == row[0].cmp(&row[1])
@@ -38,13 +38,13 @@ pub fn problem_2(data: &str) -> i32 {
 }
 
 struct Skipper<'a> {
-    row: &'a [i32],
+    row: &'a [i64],
     skip: usize,
     current: usize,
 }
 
 impl<'a> Skipper<'a> {
-    fn new(row: &'a [i32], skip: usize) -> Self {
+    fn new(row: &'a [i64], skip: usize) -> Self {
         Self {
             row,
             skip,
@@ -54,7 +54,7 @@ impl<'a> Skipper<'a> {
 }
 
 impl<'a> Iterator for Skipper<'a> {
-    type Item = i32;
+    type Item = i64;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.current == self.skip {
